# src/main/resources/openapi/dashboard-api.yaml
openapi: 3.0.3
info:
  title: CheckOn API
  version: 1.1.0
  description: |
    프론트엔드 화면 데이터 v1 연동 기준 계약이다. JSON 필드는 camelCase, 날짜는
    yyyy-MM-dd, Instant는 RFC 3339 date-time이며 서비스 논리 시간대는 Asia/Seoul이다.
    기본적으로 Bearer access token과 TEACHER 권한이 필요하다. 개발 프로필에서는 보호 API 전체를
    인증 헤더 없이 사용할 수 있으며 서버에 고정된 테스트 teacherProfileId를 사용한다.
    그 외 테넌트는 인증 principal의 teacherProfileId로 결정하며 teacherId/tenantId는 신뢰하지 않는다.
    통계, observing, feedbackGiven, 상담 일정, 문의와 Reminder 알림 발송은 범위 밖이다.

    브라우저 연동: 서버가 명시적으로 허용한 Origin만 API를 호출할 수 있다. Refresh Token은
    HttpOnly 쿠키로만 오가므로 로그인·갱신·로그아웃 요청은 fetch의 credentials: 'include' 또는
    axios의 withCredentials: true를 사용한다. Access Token은 응답 본문에서 읽어 Authorization
    Bearer 헤더로 전달한다. 202 응답의 Location 헤더는 브라우저에 노출된다.
servers:
  - url: /api/v1
    description: 환경별 호스트 뒤에 붙는 API base path (Apidog 환경에서 호스트를 설정)
security:
  - bearerAuth: []
...
paths:
...
  /problem-requests:
    post:
      tags: [Problem Generation]
      operationId: createProblemGenerationRequest
      summary: AI 문제 출제 요청
      description: |
        v1은 teacher_manual + manualTargets, language, 객관식, 지문 없음만 지원한다.
        요청과 Kafka Outbox를 같은 DB 트랜잭션에 저장하며, 동일 멱등 키와 동일
        payload는 기존 요청을 반환한다. AI 서버에는 내부 UUID와 실명을 전달하지 않는다.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProblemGenerationRequest'}
            example:
              targetKind: STUDENT
              targetId: 0198f000-0000-7000-8000-000000000003
              manualTargets: [skill.grammar.001]
              taxonomyVersion: '2026.08'
              typeTags: [CONCEPT]
              count: 3
              requestedDifficulty: MEDIUM
      responses:
        '202':
          description: 요청 저장 성공 또는 동일 요청 재사용
          headers:
            Location:
              description: 백엔드가 영속화한 출제 요청 조회 경로
              schema: {type: string, format: uri-reference}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProblemGenerationResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404':
          description: 없거나 다른 테넌트에 속한 학생 또는 클래스
          content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}, example: {code: PROBLEM_TARGET_NOT_FOUND, message: 출제 대상을 찾을 수 없습니다.}}}
        '409':
          description: 같은 멱등 키가 다른 payload에 사용됨
          content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}, example: {code: IDEMPOTENCY_CONFLICT, message: 같은 멱등 키가 다른 요청에 이미 사용되었습니다.}}}
...
  /problem-studio/students:
    get:
      tags: [Problem Studio]
      operationId: listProblemStudioStudents
      summary: Step 1 대상 학생 목록
      description: ACTIVE 학생의 화면 표시 정보와 최근 30일 이상 신호 수를 페이지로 조회한다.
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/Size'}]
      responses:
        '200': {description: 대상 학생 목록, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioStudentPage'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /problem-studio/students/{studentId}/weakness-analysis:
    get:
      tags: [Problem Studio]
      operationId: getProblemStudioWeaknessAnalysis
      summary: Step 1 최근 8주 영역·유형 분석
      parameters: [{$ref: '#/components/parameters/StudentId'}]
      responses:
        '200': {description: 최근 8주 분석, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioWeaknessAnalysis'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: 없거나 접근할 수 없는 학생, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /problem-studio/requests:
    post:
      tags: [Problem Studio]
      operationId: createProblemStudioRequest
      summary: Step 2 영역별 출제 조건으로 생성 요청
      description: 영역×유형별 문항 수의 합계는 1~20이며 요청과 Kafka Outbox를 원자 저장한다.
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioRequest'}}}
      responses:
        '202':
          description: 요청 저장
          headers:
            Location:
              description: 공통 출제 요청 상태 조회 경로
              schema: {type: string, format: uri-reference}
          content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioRequestResponse'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: 없거나 접근할 수 없는 학생, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '409': {description: 멱등 키 충돌, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /problem-studio/requests/{requestId}/review:
    get:
      tags: [Problem Studio]
      operationId: getProblemStudioReview
      summary: Step 3 AI 생성 문항 검토 조회
      parameters: [{$ref: '#/components/parameters/ProblemRequestId'}]
      responses:
        '200': {description: 문항과 검증 상태 집계, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioReview'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: 없거나 접근할 수 없는 요청, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /problem-studio/requests/{requestId}/executions/{executionId}/slots/{slotIndex}/revisions:
    post:
      tags: [Problem Studio]
      operationId: reviseProblemStudioSlot
      summary: Step 3 AI 문항 수정 요청
      description: 저장 전 슬롯에 대해 ai_refine 수정만 허용하며 baseRevisionNo로 낙관적 동시성을 검증한다.
      parameters:
        - {$ref: '#/components/parameters/ProblemRequestId'}
        - {name: executionId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: slotIndex, in: path, required: true, schema: {type: integer, minimum: 0}}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioRevisionRequest'}}}}
      responses:
        '202':
          description: 수정 요청 접수
          headers:
            Location: {description: 수정 결과가 반영되는 문제 검토 조회 경로, schema: {type: string, format: uri-reference}}
          content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioRevisionResponse'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: 요청·실행·슬롯 없음, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '409': {description: stale revision 또는 진행 중 수정 충돌, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /problem-studio/requests/{requestId}/selection:
    put:
      tags: [Problem Studio]
      operationId: replaceProblemStudioSelection
      summary: Step 4 선택 문항 교체
      parameters: [{$ref: '#/components/parameters/ProblemRequestId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioSelectionRequest'}}}}
      responses:
        '200': {description: 선택이 반영된 검토 모델, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioReview'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: 요청 없음, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '409': {description: 저장 이후 선택 변경 등 상태 충돌, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /problem-studio/requests/{requestId}/save:
    post:
      tags: [Problem Studio]
      operationId: saveProblemStudioSet
      summary: Step 4 선택 문항 저장
      parameters: [{$ref: '#/components/parameters/ProblemRequestId'}]
      responses:
        '200': {description: 동일 요청은 같은 저장 세트 반환, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioSavedSet'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: 요청 없음, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '409': {description: 저장 불가능 상태, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /problem-studio/requests/{requestId}/publish:
    post:
      tags: [Problem Studio]
      operationId: publishProblemStudioAssignment
      summary: Step 4 대상 학생에게 과제 발행
      parameters: [{$ref: '#/components/parameters/ProblemRequestId'}]
      responses:
        '200': {description: 동일 요청은 같은 발행 과제 반환, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioAssignment'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: 요청 없음, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '409': {description: 발행 불가능 상태, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /problem-studio/requests/{requestId}/printable:
    get:
      tags: [Problem Studio]
      operationId: getProblemStudioPrintableData
      summary: Step 4 프론트 PDF 렌더링용 데이터
      parameters: [{$ref: '#/components/parameters/ProblemRequestId'}]
      responses:
        '200': {description: 선택 문항과 대상 학생 인쇄 데이터, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemStudioPrintable'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: 요청 없음, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '409': {description: 선택 문항 없음, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    refreshCookie:
      type: apiKey
      in: cookie
      name: CHECKON_REFRESH
      description: 실제 쿠키 이름은 서버의 REFRESH_COOKIE_NAME 설정으로 변경할 수 있다.
  parameters:
...
    StudentId: {name: studentId, in: path, required: true, schema: {type: string, format: uuid}}
...
    ProblemRequestId: {name: requestId, in: path, required: true, schema: {type: string, format: uuid}}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 동일 논리 요청의 재시도에서는 같은 값을 사용한다.
      schema: {type: string, minLength: 8, maxLength: 200, pattern: '^[A-Za-z0-9._:-]+$'}
...
    Page: {name: page, in: query, required: false, description: 0부터 시작, schema: {type: integer, format: int32, minimum: 0, default: 0}}
    Size: {name: size, in: query, required: false, schema: {type: integer, format: int32, minimum: 1, maximum: 100, default: 20}}
  responses:
    Unauthorized:
      description: Access Token이 없거나 유효하지 않음
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}, example: {code: UNAUTHORIZED, message: 인증이 필요합니다.}}}
    Forbidden:
      description: 인증됐지만 TEACHER 권한이 아님
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}, example: {code: FORBIDDEN, message: 요청한 작업을 수행할 권한이 없습니다.}}}
    InvalidRequest:
      description: 요청 값 오류
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}, example: {code: INVALID_REQUEST, message: 요청 값을 확인해 주세요.}}}
...
  schemas:
...
    PaginationMetadata:
      type: object
      required: [pageNumber, pageSize, itemCount, totalItemCount, totalPageCount, isFirst, isLast]
      properties:
        pageNumber: {type: integer, format: int32, minimum: 0}
        pageSize: {type: integer, format: int32, minimum: 1, maximum: 100}
        itemCount: {type: integer, format: int32, minimum: 0}
        totalItemCount: {type: integer, format: int64, minimum: 0}
        totalPageCount: {type: integer, format: int64, minimum: 0}
        isFirst: {type: boolean}
        isLast: {type: boolean}
...
    ProblemTargetKind: {type: string, enum: [STUDENT, CLASS]}
    ProblemTypeTag: {type: string, enum: [FACT, INFER, CRITIC, CONCEPT]}
    ProblemDifficulty: {type: string, enum: [LOW, MEDIUM, HIGH]}
    ProblemGenerationStatus:
      type: string
      enum: [QUEUED, DISPATCHED, RUNNING, SUCCEEDED, PARTIAL_SUCCESS, FAILED, CANCELLED, DELIVERY_FAILED]
    ProblemGenerationRequest:
      type: object
      additionalProperties: false
      required: [targetKind, targetId, manualTargets, taxonomyVersion, typeTags, count]
      properties:
        targetKind: {$ref: '#/components/schemas/ProblemTargetKind'}
        targetId: {type: string, format: uuid}
        manualTargets:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 120, pattern: '^[a-zA-Z0-9][a-zA-Z0-9._:-]*$'}
        taxonomyVersion: {type: string, minLength: 1, maxLength: 80}
        typeTags:
          type: array
          minItems: 1
          maxItems: 4
          uniqueItems: true
          items: {$ref: '#/components/schemas/ProblemTypeTag'}
        count: {type: integer, format: int32, minimum: 1, maximum: 10}
        requestedDifficulty:
          type: string
          allOf: [{$ref: '#/components/schemas/ProblemDifficulty'}]
          nullable: true
    ProblemGenerationResponse:
      type: object
      required:
        - requestId
        - targetKind
        - status
        - requestedAt
        - replayed
      properties:
        requestId: {type: string, format: uuid}
        targetKind: {$ref: '#/components/schemas/ProblemTargetKind'}
        status: {$ref: '#/components/schemas/ProblemGenerationStatus'}
        jobId: {type: string, nullable: true}
        executionId: {type: string, nullable: true}
        setId: {type: string, nullable: true}
        resultStatus: {type: string, nullable: true}
        errorCode: {type: string, nullable: true}
        result: {type: object, nullable: true, additionalProperties: true}
        versions: {type: object, nullable: true, additionalProperties: true}
        requestedAt: {type: string, format: date-time}
        dispatchedAt: {type: string, format: date-time, nullable: true}
        completedAt: {type: string, format: date-time, nullable: true}
        replayed: {type: boolean, description: 동일 멱등 키와 동일 payload로 기존 요청을 재사용했는지 여부}
    ProblemStudioStudentPage:
      type: object
      required: [metadata, items]
      properties:
        metadata: {$ref: '#/components/schemas/PaginationMetadata'}
        items: {type: array, items: {$ref: '#/components/schemas/ProblemStudioStudent'}}
    ProblemStudioStudent:
      type: object
      required: [studentId, studentName, managedDays, recentSignalCount]
      properties:
        studentId: {type: string, format: uuid}
        studentName: {type: string}
        className: {type: string, nullable: true}
        subject: {type: string, nullable: true}
        managedDays: {type: integer, format: int64, minimum: 1}
        recentSignalCount: {type: integer, minimum: 0}
        notableSummary: {type: string, nullable: true}
    ProblemStudioWeaknessAnalysis:
      type: object
      required: [diagnosisId, studentId, windowStart, windowEnd, minimumSampleSize, cells, skillNodes, generationCapabilities, diagnosis]
      properties:
        diagnosisId: {type: string, format: uuid, description: Step 2 요청이 그대로 참조해야 하는 불변 진단 ID}
        studentId: {type: string, format: uuid}
        windowStart: {type: string, format: date}
        windowEnd: {type: string, format: date}
        minimumSampleSize: {type: integer, enum: [10]}
        studentAveragePercent: {type: number, nullable: true}
        cells: {type: array, items: {$ref: '#/components/schemas/ProblemStudioWeaknessCell'}}
        skillNodes:
          type: array
          description: 교사가 출제 요청에서 명시적으로 선택할 수 있는 진단 근거 노드다.
          items: {$ref: '#/components/schemas/ProblemStudioSkillNodeCandidate'}
        generationCapabilities:
          type: array
          description: 현재 AI evidence로 생성 가능한 셀만 포함한다. 프론트는 이 목록에 없는 셀을 비활성화한다.
          items: {$ref: '#/components/schemas/ProblemStudioGenerationCapability'}
        diagnosis: {$ref: '#/components/schemas/ProblemStudioDiagnosisProvenance'}
    ProblemStudioDiagnosisProvenance:
      type: object
      required: [status, snapshotHash]
      properties:
        status: {type: string, enum: [GENERATED, REJECTED_INSUFFICIENT, UNAVAILABLE]}
        statusReason: {type: string, nullable: true}
        snapshotHash: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
        taxonomyVersion: {type: string, nullable: true}
        graphVersion: {type: string, nullable: true}
        configVersion: {type: string, nullable: true}
    ProblemStudioGenerationCapability:
      type: object
      required: [areaTag, typeTag, maximumCount, recommendedMaximumCount]
      properties:
        areaTag: {type: string, enum: [language, reading, literature, speech_writing, media]}
        typeTag: {type: string, enum: [FACT, INFER, CRITIC, CONCEPT]}
        maximumCount: {type: integer, enum: [20]}
        recommendedMaximumCount: {type: integer, enum: [3]}
    ProblemStudioSkillNodeCandidate:
      type: object
      required: [skillNodeId, verdict, basis]
      properties:
        skillNodeId: {type: string, maxLength: 120}
        verdict: {type: string, description: ok가 아닌 진단 노드만 출제 대상으로 선택할 수 있다.}
        basis: {type: array, items: {type: string}}
    ProblemStudioWeaknessCell:
      type: object
      required: [areaTag, typeTag, solvedCount, correctCount, evaluation]
      properties:
        areaTag: {type: string}
        typeTag: {type: string, enum: [FACT, INFER, CRITIC, CONCEPT]}
        solvedCount: {type: integer, minimum: 0}
        correctCount: {type: integer, minimum: 0}
        accuracyPercent: {type: number, nullable: true}
        gapFromStudentAveragePercent: {type: number, nullable: true}
        evaluation: {type: string, enum: [GOOD, WEAK_SIGNAL, WEAK_CONFIRMED, ON_HOLD]}
    ProblemStudioRequest:
      type: object
      additionalProperties: false
      required: [studentId, diagnosisId, targets, difficulty]
      properties:
        studentId: {type: string, format: uuid}
        diagnosisId: {type: string, format: uuid, description: 동일 학생의 GENERATED 진단 ID}
        targets:
          type: array
          minItems: 1
          maxItems: 20
          items: {$ref: '#/components/schemas/ProblemStudioTarget'}
        difficulty: {$ref: '#/components/schemas/ProblemDifficulty'}
    ProblemStudioTarget:
      type: object
      additionalProperties: false
      required: [areaTag, typeTag, count, skillNodeId]
      properties:
        areaTag: {type: string, enum: [language, reading, literature, speech_writing, media]}
        typeTag: {type: string, enum: [FACT, INFER, CRITIC, CONCEPT]}
        count: {type: integer, minimum: 1, maximum: 20}
        skillNodeId: {type: string, maxLength: 120}
        passage: {$ref: '#/components/schemas/ProblemStudioPassage'}
        workSelection: {$ref: '#/components/schemas/ProblemStudioWorkSelection'}
    ProblemStudioPassage:
      type: object
      additionalProperties: false
      description: reading·speech_writing·media 영역에서 필수이며 language·literature에서는 허용하지 않는다.
      properties:
        areaTag: {type: string, maxLength: 40}
        domain: {type: string, maxLength: 40}
        topicHint: {type: string, maxLength: 300}
        wordCount: {type: integer, minimum: 1, maximum: 5000}
        sentenceComplexity: {type: string, maxLength: 20}
        paragraphCount: {type: integer, minimum: 2, maximum: 6}
        sourceKind: {type: string, maxLength: 40}
        bannedTopicsVersion: {type: string, enum: [pg-banned-v1]}
    ProblemStudioWorkSelection:
      type: object
      additionalProperties: false
      description: literature 영역에서만 필수다.
      properties:
        genre: {type: string, maxLength: 40}
        era: {type: string, maxLength: 100}
        conceptKeywords: {type: array, maxItems: 20, items: {type: string, maxLength: 100}}
    ProblemStudioRequestResponse:
      type: object
      required: [requestId, status, replayed]
      properties:
        requestId: {type: string, format: uuid}
        status: {$ref: '#/components/schemas/ProblemGenerationStatus'}
        replayed: {type: boolean}
    ProblemValidationStatus: {type: string, enum: [PASSED, REVIEW_REQUIRED, UNVERIFIABLE, EXCLUDED]}
    ProblemStudioReview:
      type: object
      required: [requestId, requestStatus, projectionStatus, counts, items, slots]
      properties:
        requestId: {type: string, format: uuid}
        requestStatus: {$ref: '#/components/schemas/ProblemGenerationStatus'}
        projectionStatus: {type: string, enum: [PENDING, PROJECTED, PARTIAL, UNSUPPORTED]}
        projectionErrorCode: {type: string, nullable: true}
        counts: {$ref: '#/components/schemas/ProblemStudioReviewCounts'}
        items: {type: array, items: {$ref: '#/components/schemas/ProblemStudioReviewItem'}}
        slots: {type: array, items: {$ref: '#/components/schemas/ProblemStudioReviewSlot'}}
    ProblemStudioReviewSlot:
      type: object
      required: [executionId, slotIndex, status, currentRevisionNo, availableActions]
      properties:
        executionId: {type: string, format: uuid}
        slotIndex: {type: integer, minimum: 0}
        itemId: {type: string, format: uuid, nullable: true}
        externalItemId: {type: string, nullable: true}
        status: {$ref: '#/components/schemas/ProblemValidationStatus'}
        currentRevisionNo: {type: integer, minimum: 0}
        availableActions: {type: array, uniqueItems: true, items: {type: string, enum: [refine]}}
        reviewReason: {type: string, nullable: true}
        failureReason: {type: string, nullable: true}
    ProblemStudioReviewCounts:
      type: object
      required: [passed, reviewRequired, unverifiable, excluded]
      properties:
        passed: {type: integer, minimum: 0}
        reviewRequired: {type: integer, minimum: 0}
        unverifiable: {type: integer, minimum: 0}
        excluded: {type: integer, minimum: 0}
    ProblemStudioReviewItem:
      type: object
      required: [itemId, ordinal, areaTag, typeTag, skillNodeId, correctNo, stem, options, validationStatus, selected]
      properties:
        itemId: {type: string, format: uuid}
        ordinal: {type: integer, minimum: 1}
        externalItemId: {type: string, nullable: true}
        areaTag: {type: string, enum: [language, reading, literature, speech_writing, media]}
        typeTag: {type: string, enum: [FACT, INFER, CRITIC, CONCEPT]}
        skillNodeId: {type: string}
        correctNo: {type: integer, minimum: 1, maximum: 5}
        stem: {type: string}
        passage: {type: string, nullable: true}
        options: {type: array, minItems: 2, items: {$ref: '#/components/schemas/ProblemStudioOption'}}
        correctAnswerText: {type: string, nullable: true}
        explanation: {type: string, nullable: true}
        sourceBasis: {type: string, nullable: true}
        validationStatus: {$ref: '#/components/schemas/ProblemValidationStatus'}
        validationMessage: {type: string, nullable: true}
        selected: {type: boolean}
    ProblemStudioOption:
      type: object
      required: [position, content, correct]
      properties:
        position: {type: integer, minimum: 1}
        content: {type: string}
        correct: {type: boolean}
        whyWrong: {type: string, nullable: true}
        misconceptionTag: {type: string, nullable: true}
    ProblemStudioRevisionRequest:
      type: object
      additionalProperties: false
      required: [baseRevisionNo, revisionKind, instruction]
      properties:
        baseRevisionNo: {type: integer, minimum: 0}
        revisionKind: {type: string, enum: [ai_refine]}
        instruction: {type: string, minLength: 1, maxLength: 2000}
    ProblemStudioRevisionResponse:
      type: object
      required: [revisionRequestId, status, replayed]
      properties:
        revisionRequestId: {type: string, format: uuid}
        status: {type: string, enum: [PENDING, DISPATCHED, COMPLETED, FAILED]}
        replayed: {type: boolean}
    ProblemStudioSelectionRequest:
      type: object
      additionalProperties: false
      required: [itemIds]
      properties:
        itemIds: {type: array, maxItems: 20, uniqueItems: true, items: {type: string, format: uuid}}
    ProblemStudioSavedSet:
      type: object
      required: [problemSetId, itemCount, status, savedAt]
      properties:
        problemSetId: {type: string, format: uuid}
        itemCount: {type: integer, minimum: 1}
        status: {type: string, enum: [SAVED, PUBLISHED]}
        savedAt: {type: string, format: date-time}
    ProblemStudioAssignment:
      type: object
      required: [assignmentId, problemSetId, studentId, itemCount, status, publishedAt]
      properties:
        assignmentId: {type: string, format: uuid}
        problemSetId: {type: string, format: uuid}
        studentId: {type: string, format: uuid}
        itemCount: {type: integer, minimum: 1}
        status: {type: string, enum: [PUBLISHED]}
        publishedAt: {type: string, format: date-time}
    ProblemStudioPrintable:
      type: object
      required: [requestId, studentId, studentName, generatedAt, items]
      properties:
        requestId: {type: string, format: uuid}
        studentId: {type: string, format: uuid}
        studentName: {type: string}
        className: {type: string, nullable: true}
        subject: {type: string, nullable: true}
        generatedAt: {type: string, format: date-time}
        items: {type: array, minItems: 1, items: {$ref: '#/components/schemas/ProblemStudioReviewItem'}}
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
        detail:
          type: object
          nullable: true
          properties:
            reason: {type: string, nullable: true}
            currentRevisionNo: {type: integer, minimum: 0, nullable: true}